+++ src/burp/controller/TriggerParamGuesser.java
@@ -4,8 +4,8 @@ import burp.IHttpRequestResponse;
 import burp.IHttpService;
 import burp.IParameter;
 import burp.IResponseInfo;
-import burp.ParamGrabber;
-import burp.ParamGuesser;
+import burp.model.param.ParamGrabber;
+import burp.model.param.ParamGuesser;
 import burp.model.utilities.Utilities;
 import burp.view.ConfigurableSettings;
 import org.apache.commons.collections4.queue.CircularFifoQueue;
